// src/payment/interfaces/IMintCoin.ts
// Generated by https://quicktype.io

export interface IMintCoin {
  chainId:              number;
  hash:                 string;
  transactionId:        string;
  value:                GasLimit;
  gasLimit:             GasLimit;
  to:                   string;
  from:                 string;
  data:                 string;
  nonce:                number;
  status:               string;
  speed:                string;
  validUntil:           string;
  createdAt:            string;
  sentAt:               string;
  pricedAt:             string;
  isPrivate:            boolean;
  maxFeePerGas:         GasLimit;
  maxPriorityFeePerGas: GasLimit;
}

export interface GasLimit {
  type: string;
  hex:  string;
}